release: sincroniza código de produção da dev
- feat(type,viewport): implementar operadores multiplicativos em Scale e propriedade zoom
- refactor(layout): exigir parametro ref em ViewportLayoutStrategy.fit
- feat(layout): implementar ViewportLayoutStrategy e tipagem LayoutRef
- refactor(buffer): substituir zarr por mmapbuffer para paginacao out-of-core

sync-point: 7f9e484

+++ src/anicrop/interfaces/layout.py
@@ -25,8 +25,8 @@ class LayoutStrategy(ABC):
     @abstractmethod
     def resize_bounds(
         self,
-        new_width: int,
-        new_height: int,
+        new_width: float,
+        new_height: float,
         anchor_x: float = 0.5,
         anchor_y: float = 0.5,
     ) -> bool:
